layouts/blog: Use .Summary in blog list
This is not ideal, and we may end up reverting this commit. But I prefer
to use the .Summary parameter in the blog post list rather than the
plain contents truncated to 250 characters. This allows us to specify
the summary explicitly [0], and to avoid having the text and byline for
the featured picture for the post appearing in the summary.

The diff with Docsy's version is minimal:

    --- _vendor/github.com/google/docsy/layouts/blog/list.html	2025-03-19 23:55:44.046202993 +0000
    +++ layouts/blog/list.html	2025-03-19 22:40:00.590083142 +0000
    @@ -23,7 +23,7 @@
                 {{ end -}}
               </header>
               {{ partial "featured-image.html" (dict "p" . "w" 250 "h" 125 "class" "float-start me-3 pt-1 d-none d-md-block") -}}
    -          <p class="pt-0 mt-0">{{ .Plain | safeHTML | truncate 250 }}</p>
    +          <p class="pt-0 mt-0">{{ .Summary }}</p>
               <p class="pt-0"><a href="{{ .RelPermalink }}" aria-label="{{ T "ui_read_more"}} - {{ .LinkTitle }}">{{ T "ui_read_more"}}</a></p>
             </div>
           </li>

[0] https://gohugo.io/content-management/summaries/