index: Rework list of actions for DGRA
Improve the list of actions displayed on the home page. This is an
important list, because this is the first thing we see as we scroll down
from the title on the home page of the site. Remove the "-ing"
terminations, bring forward the support/connect aspects. Try to replace
words that sound vague ("useful info" -> "practical info", that's not a
big improvement but still sounds a bit less hand-waiving to my ear.