docs/contact: Add a contact page in the docs, as practical info
This page is not for contacting the DGRA (although we still link to our
own contact page), but instead to find contact details for the different
people or entities whom it might be relevant to get in touch with about
Darwin Green. Mention Barratt, the Community Development Team,
authorities for anti-social behaviour, etc.

The page is not complete yet, in particular we need to provide a contact
point for Barratt.