lore/names: Add Sedgwick, Jenyns, remove Fitzroy
The street sign for Fitzroy Crescent has been replaced with a new sign
indicating Sedgwick Crescent, so let's update our list. We can add
Fitzroy again in the future if it turns out Fitzroy Crescent was to be a
thing after all. Also add an entry for the upcoming Jenyns Drive.