blog: Add tags and categories
We have enough posts that it starts to make sense to sort them with tag
and categories, especially for finding the Community Forum reports, for
example.

Deactivate the "taxonomy cloud" in the right sidebar. From the docs [0],
it looks like we should pass nothing for the parameter rather than this
"empty" value with the dash, but I couldn't get it to work (while the
dash does).

[0]: https://www.docsy.dev/docs/adding-content/taxonomy/#parameters